Stopwatch: adding an extra double cast just to be sure.
git-svn-id: http://dev.outerthought.org/svn_public/outerthought_lilyproject/trunk@4553 b80d95f8-e103-0410-bd55-e03a9dc29394

+++ global/testclient-fw/src/main/java/org/lilyproject/testclientfw/Stopwatch.java
@@ -32,7 +32,7 @@ public class Stopwatch {
             LogFactory.getLog(Stopwatch.class).error("Stopwatch.stop() called but it was not started.");
         }
 
-        double duration = (System.nanoTime() - ((double)data.startedAt)) / 1e6;
+        double duration = (((double)System.nanoTime()) - ((double)data.startedAt)) / 1e6;
         data.started = false;
         return duration;
     }
